# jwe/CompactSerialize: improve performance.