Use concatenate instead of `fmt.Sprintf(...)`. Reduce memory allocation.

on my machine: m1 pro, 32gb ram.

```
name                 old time/op    new time/op    delta
CompactSerialize-10    1.32µs ± 1%    1.12µs ± 1%  -15.40%  (p=0.000 n=19+19)

name                 old alloc/op   new alloc/op   delta
CompactSerialize-10    2.21kB ± 0%    2.13kB ± 0%   -3.62%  (p=0.000 n=20+20)

name                 old allocs/op  new allocs/op  delta
CompactSerialize-10      25.0 ± 0%      20.0 ± 0%  -20.00%  (p=0.000 n=20+20)
```