@jsha 2023-11-15
This is a neat finding! Do you have an impression of why concatenation is faster than Sprintf in this case?

If string allocations are a key factor in the hot path here, I would expect that it would be faster still to do this:

 - calculate the encoded length of each component (`base64` provides `EncodedLen`), plus the separator dot characters.
 - allocate a `[]byte` of that length.
 - write the encoded values directly into that `[]byte` using `Encode`.
 - also write the dots.

Would you be interested in benchmarking that variant as well?

Also it looks like we follow the same pattern for JSONWebSignature.CompactSerialize. Whichever approach we wind up with here, would you apply it there too?