@Lasiar 2023-11-16
Hi @jsha, thank you for feedback!

About string formatter, [here](https://forum.golangbridge.org/t/fmt-sprintf-vs-string-concatenation/23006) is an explanation.

About use preallocating slice.
Yes, this is faster on the same machine (m1 pro, 32gb ram):

```
name                 time/op
CompactSerialize-10   848ns ±11%

name                 alloc/op
CompactSerialize-10  1.63kB ± 0%

name                 allocs/op
CompactSerialize-10    11.0 ± 0% 
```

But the code is more unclear:
<details>
<summary>example</summary>


encoding.go:
```go
func base64URLEncodeLen(sl []byte) int {
	return base64.RawURLEncoding.EncodedLen(len(sl))
}

// base64URLEncodeAppend appends base64-encoded src to dst and returns the dst.
func base64URLEncodeAppend(dst, src []byte) []byte {
	old := len(dst)
	need := base64URLEncodeLen(src)
	end := old + need
	for cap(dst)-len(dst) < need {
		dst = append(dst[:cap(dst)], 0)
	}

	base64.RawURLEncoding.Encode(dst[old:end], src)
	return dst[:end]
}
```

jwe.go:
```go
// CompactSerialize serializes an object using the compact serialization format.
func (obj JSONWebEncryption) CompactSerialize() (string, error) {
	if len(obj.recipients) != 1 || obj.unprotected != nil ||
		obj.protected == nil || obj.recipients[0].header != nil {
		return "", ErrNotSupported
	}

	serializedProtected := mustSerializeJSON(obj.protected)

	const countOfDots = 4
	lenSerialized := base64URLEncodeLen(serializedProtected) +
		base64URLEncodeLen(obj.recipients[0].encryptedKey) +
		base64URLEncodeLen(obj.iv) +
		base64URLEncodeLen(obj.ciphertext) +
		base64URLEncodeLen(obj.tag) +
		countOfDots

	out := make([]byte, 0, lenSerialized)

	out = base64URLEncodeAppend(out, serializedProtected)
	out = append(out, '.')

	out = base64URLEncodeAppend(out, obj.recipients[0].encryptedKey)
	out = append(out, '.')

	out = base64URLEncodeAppend(out, obj.iv)
	out = append(out, '.')

	out = base64URLEncodeAppend(out, obj.ciphertext)
	out = append(out, '.')

	out = base64URLEncodeAppend(out, obj.tag)

	return string(out), nil
}
```

jws.go

```go
func (obj JSONWebSignature) compactSerialize(detached bool) (string, error) {
	if len(obj.Signatures) != 1 || obj.Signatures[0].header != nil || obj.Signatures[0].protected == nil {
		return "", ErrNotSupported
	}

	serializedProtected := mustSerializeJSON(obj.Signatures[0].protected)
	const countOfDots = 3

	maxLenSerialized := base64URLEncodeLen(serializedProtected) +
		base64URLEncodeLen(obj.payload) +
		base64URLEncodeLen(obj.Signatures[0].Signature) +
		countOfDots

	serialized := make([]byte, 0, maxLenSerialized)

	serialized = base64URLEncodeAppend(serialized, serializedProtected)
	serialized = append(serialized, '.')

	if !detached {
		serialized = base64URLEncodeAppend(serialized, obj.payload)
	}
	serialized = append(serialized, '.')

	serialized = base64URLEncodeAppend(serialized, obj.Signatures[0].Signature)

	return string(serialized), nil
}
```
</details>

If you accept these changes, I can push.