@jsha 2023-11-17
Thanks for exploring the alternate approach. That's a nice significant speedup.

It seems like we can make it a little clearer by factoring out this logic into `joinWithDots(inputs... []byte)`, which would work for any number of inputs and could be used by both call sites.

Also, I like the safety of your approach where the slice is made with cap N and len 0, and appended to. But this feels a little awkward:

```go
	for cap(dst)-len(dst) < need {
		dst = append(dst[:cap(dst)], 0)
	}
```

I think we would be better off initializing the slice with cap N and len N, and keeping track of the actually-initialized part ourselves. That might also save a little on that loop, depending on how the optimizer handles it.